Refactor deserializedModel
